AI_Employee_Vault/requirements.txt
# AI Employee - Python Dependencies
# Silver Tier Requirements

# Core dependencies (required)
watchdog>=3.0.0          # File system monitoring
playwright>=1.40.0       # Browser automation (WhatsApp, LinkedIn)

# Gmail integration (required for Email MCP and Gmail Watcher)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Utilities
python-dotenv>=1.0.0     # Environment variable management
python-dateutil>=2.8.0   # Date utilities (CEO Briefing)

# Development
pytest>=7.0.0            # Testing (optional)
